made token optional in Genius, moved Artist and Song to types added TokenRequiredError updated docs,
