+++ README.md
@@ -11,7 +11,17 @@
 `lyricsgenius`'s full documentation is online at [Read the Docs](https://lyricsgenius.readthedocs.io/en/master/).
 
 ## Setup
-Before using this package you'll need to sign up for a (free) account that authorizes access to [the Genius API](http://genius.com/api-clients). The Genius account provides a `access_token` that is required by the package. See the [Usage section](https://github.com/johnwmillr/LyricsGenius#usage) below for examples.
+Genius's API has two interaces: the developer's API and the tokenfree, undocumented public API. The developer's API has limited endpoints and needs a token (client- or user access token). On the other hand the public API is not limited and needs no token to work.
+There are three ways to use the package:
+* **With a user access token**: User access tokens provide the most functionality but they are really useful if you're planning to use Genius's [Web Annotator](https://genius.com/web-annotator). If you're not planning to use it (or don't even know what it is), that leaves you with using a client access token or no tokens.
+* **With a client access token**: With a client access token you can get information using both the developers and the public API.
+* **Without an access token**: Without an access token you won't be able to get information using the developers API, but the public API -which doesn't need a token- covers all of the developers API endpoints and many more! So you'll be okay.
+
+Here's what you need to do for each approach:
+- User Access Tokens: You need to sign up/log in on Genius, [create an app](http://genius.com/api-clients) and get a user token using OAuth2 (our [OAuth2](https://lyricsgenius.readthedocs.io/en/latest/reference/auth.html#auth) class can help you with that).
+- Client Access Tokens: You need to sign up/log in on Genius and [create an app](http://genius.com/api-clients). Then on your app's page, click *GENERATE CLIENT ACCESS TOKEN* and you'll get a client access token.
+- No Tokens: No need to do anything.
+
 
 ## Installation
 `lyricsgenius` requires Python 3.
@@ -33,7 +43,7 @@ Import the package and search for songs by a given artist:
 
 ```python
 import lyricsgenius
-genius = lyricsgenius.Genius("my_access_token_here")
+genius = lyricsgenius.Genius() # or Genius(token)
 artist = genius.search_artist("Andy Shauf", max_songs=3, sort="title")
 print(artist.songs)
 ```
@@ -49,7 +59,7 @@ print(artist.songs)
 Search for a single song by the same artist:
 
 ```python
-song = genius.search_song("To You", artist.name)
+song = artist.song("To You")
 print(song.lyrics)
 ```
 
@@ -57,6 +67,8 @@ Add the song to the artist object:
 
 ```python
 artist.add_song(song)
+# the Artist object also accepts song names:
+# artist.add_song("To You")
 ```
 
 Save the artist's songs to a JSON file:
@@ -76,8 +88,13 @@ genius.excluded_terms = ["(Remix)", "(Live)"] # Exclude songs with these words i
 
 You can also call the package from the command line:
 
+With a token:
 ```bash
 export GENIUS_ACCESS_TOKEN="my_access_token_here"
+python3 -m lyricsgenius --help
+```
+Without a token:
+```bash
 python3 -m lyricsgenius --help
 ```
