+++ docs/src/index.rst
@@ -52,12 +52,11 @@ file:
 
    from lyricsgenius import Genius
 
-   genius = Genius(token)
+   genius = Genius(token)  # the token is optional
    genius.search_artist('Andy Shauf')
    artist.save_lyrics()
 
-But before using the library you will need to get an access token. Head over
-to :ref:`setup` to get started.
+Head over to :ref:`setup` to get started.
 
 .. toctree::
    :maxdepth: 1
