+++ docs/src/reference.rst
@@ -6,13 +6,12 @@ Reference
 All public objects are documented in these sections.
 
 ================== =============================
-:ref:`sender`      Sender
-:ref:`auth`        OAuth2
-:ref:`api`         API
-:ref:`public_api`  Public API
-:ref:`Genius`      Genius
-:ref:`Artist`	    Artist object
-:ref:`Song`        Song object
+:ref:`api`         API and PublicAPI classes
+:ref:`auth`        OAuth2 class
+:ref:`exceptions`  Exceptions
+:ref:`Genius`      Genius class
+:ref:`sender`      Request sender
+:ref:`types`	    Types
 :ref:`utils`       Utility functions
 ================== =============================
 
@@ -20,12 +19,6 @@ All public objects are documented in these sections.
    :maxdepth: 2
    :caption: Index
    :hidden:
+   :glob:
 
-   reference/sender
-   reference/auth
-   reference/api
-   reference/public_api
-   reference/genius
-   reference/artist
-   reference/song
-   reference/utils
+   reference/*
