--- docs/src/reference/artist.rst
@@ -1,37 +1,0 @@
-.. _artist:
-.. currentmodule:: lyricsgenius.artist
-.. toctree::
-   :maxdepth: 2
-   :hidden:
-   :caption: Artist
-
-Artist
-=======================
-The Artist object which holds the details of the artist
-and the :ref:`song` objects of that artist.
-
-Properties
------------
-.. autosummary::
-   :nosignatures:
-
-   Artist.name
-   Artist.image_url
-   Artist.songs
-
-
-Methods
---------
-.. autosummary::
-   :nosignatures:
-
-   Artist.song
-   Artist.add_song
-   Artist.to_json
-   Artist.to_text
-   Artist.save_lyrics
-
-.. autoclass:: lyricsgenius.artist.Artist
-    :members:
-    :member-order: bysource
-    :no-show-inheritance:
