+++ docs/src/reference/exceptions.rst
@@ -1,0 +1,14 @@
+.. _exceptions:
+.. currentmodule:: lyricsgenius.exceptions
+.. toctree::
+   :maxdepth: 2
+   :hidden:
+   :caption: Exceptions
+
+Exceptions
+==========
+Package-defined exceptions. Visit :ref:`other_guides` to see how
+LyricsGenius handles HTTP errors. 
+
+.. autoclass:: TokenRequiredError
+   :show-inheritance:
