--- docs/src/reference/song.rst
@@ -1,45 +1,0 @@
-.. _song:
-.. currentmodule:: lyricsgenius.song
-.. toctree::
-   :maxdepth: 2
-   :hidden:
-   :caption: Song
-
-Song
-=======================
-This is the Song object which holds the details of the song.
-
-
-Properties
------------
-.. autosummary::
-   :nosignatures:
-
-   Song.title
-   Song.artist
-   Song.lyrics
-   Song.album
-   Song.year
-   Song.url
-   Song.album_url
-   Song.featured_artists
-   Song.producer_artists
-   Song.media
-   Song.writer_artists
-   Song.song_art_image_url
-
-
-Methods
---------
-.. autosummary::
-   :nosignatures:
-
-   Song.to_dict
-   Song.to_json
-   Song.to_text
-   Song.save_lyrics
-
-.. autoclass:: Song
-    :members:
-    :member-order: bysource
-    :no-show-inheritance:
