+++ docs/src/usage.rst
@@ -1,4 +1,5 @@
 .. _usage:
+.. currentmodule:: lyricsgenius
 
 Usage
 ==================
@@ -7,8 +8,9 @@ Import the package and search for songs by a given artist:
 
 .. code:: python
 
-   import lyricsgenius
-   genius = lyricsgenius.Genius("my_access_token_here")
+   from lyricsgenius import Genius
+
+   genius = Genius()  # or Genius(token)
    artist = genius.search_artist("Andy Shauf", max_songs=3, sort="title")
    print(artist.songs)
 
@@ -16,14 +18,23 @@ Search for a single song by the same artist:
 
 .. code:: python
 
+   # Way 1
    song = genius.search_song("To You", artist.name)
+
+   # Way 2
+   # this will search artist.songs first
+   # and if not found, uses search_song
+   song = artist.song("To You")
+
    print(song.lyrics)
 
-Add the song to the :ref:`artist` object:
+Add the song to the :class:`Artist <types.Artist>` object:
 
 .. code:: python
 
    artist.add_song(song)
+   # add_song accepts song names as well:
+   # artist.add_song("To You")
 
 Save the artist’s songs to a JSON file:
