+++ lyricsgenius/__init__.py
@@ -1,6 +1,7 @@
 # GeniusAPI
 # John W. Miller
 # See LICENSE for details
+"""A library that provides a Python interface to the Genius API"""
 
 import sys
 assert sys.version_info[0] == 3, "LyricsGenius requires Python 3."
@@ -8,7 +9,6 @@ from lyricsgenius.genius import Genius
 from lyricsgenius.api import API, PublicAPI
 from lyricsgenius.auth import OAuth2
 
-"""A library that provides a Python interface to the Genius API"""
 
 __author__ = 'John W. Miller'
 __url__ = 'https://github.com/johnwmillr/LyricsGenius'
