+++ lyricsgenius/api/__init__.py
@@ -1,1 +1,2 @@
 from .api import API, PublicAPI
+from .base import Sender, check_token
