+++ lyricsgenius/types/__init__.py
@@ -1,0 +1,2 @@
+from .artist import Artist
+from .song import Song
