+++ lyricsgenius/types/artist.py
@@ -60,7 +60,7 @@ class Artist(object):
         """Song objects saved for the artist.
 
         Returns:
-            :obj:`list`: A list contatining :class:`Song <song.Song>`
+            :obj:`list`: A list contatining :class:`Song <types.Song>`
             objects.
 
         """
@@ -85,7 +85,7 @@ class Artist(object):
         song's artist is the same as the `Artist` object.
 
         Args:
-            new_song (:class:`Song <lyricsgenius.song.Song>`): Song to be added.
+            new_song (:class:`Song <lyricsgenius.types.Song>`): Song to be added.
             verbose (:obj:`bool`, optional): prints operation result.
             include_features (:obj:`bool`, optional): If True, includes tracks
                 featuring the artist.
@@ -138,7 +138,7 @@ class Artist(object):
             sanitize (:obj:`bool`): Sanitizes the filename if `True`.
 
         Returns:
-            :obj:`Song <song.Song>` \\|‌ :obj:`None`: If it can't find the song,
+            :obj:`Song <types.Song>` \\|‌ :obj:`None`: If it can't find the song,
                 returns *None*.
 
         """
@@ -234,7 +234,7 @@ class Artist(object):
         artist songs.
         If you only want the songs lyrics, set :obj:`extension` to `txt`.
         If you choose to go with JSON (which is the default extension), you can access
-        the lyrics by accessing the :class:`Song <song.Song>`
+        the lyrics by accessing the :class:`Song <types.Song>`
         objects inside the `songs` key of the JSON file.
         Take a look at the example below.
