+++ tests/test_public_methods.py
@@ -1,6 +1,9 @@
 import unittest
 
-from .test_genius import genius
+try:
+    from .test_genius import genius
+except ModuleNotFoundError:
+    from test_genius import genius
 
 
 class TestAlbumMethods(unittest.TestCase):
